+++ R/plot.R
@@ -57,6 +57,8 @@
 #' @param alpha Alpha value for the confidence bands.
 #' @param line_size Numeric, size of the line geoms.
 #' @param dot_size Numeric, size of the point geoms.
+#' @param dot_alpha Alpha value for data points, when `show_data = TRUE`.
+#' @param dot_shape Shape of data points, when `show_data = TRUE`.
 #' @param dodge Value for offsetting or shifting error bars, to avoid overlapping.
 #'   Only applies, if a factor is plotted at the x-axis (in such cases, the
 #'   confidence bands are replaced by error bars automatically), or if
@@ -64,7 +66,6 @@
 #' @param use_theme Logical, if `TRUE`, a slightly tweaked version of ggplot's
 #'   minimal-theme, `theme_ggeffects()`, is applied to the plot. If `FALSE`, no
 #'   theme-modifications are applied.
-#' @param dot_alpha Alpha value for data points, when `show_data = TRUE`.
 #' @param jitter Numeric, between 0 and 1. If not `NULL` and `show_data = TRUE`,
 #'   adds a small amount of random variation to the location of data points dots,
 #'   to avoid overplotting. Hence the points don't reflect exact values in the
@@ -166,11 +167,12 @@ plot.ggeffects <- function(x,
                            # appearance colors and geoms
                            colors = NULL,
                            alpha = 0.15,
+                           dot_size = NULL,
                            dot_alpha = 0.35,
+                           dot_shape = NULL,
+                           line_size = NULL,
                            jitter = NULL,
                            dodge = 0.25,
-                           dot_size = NULL,
-                           line_size = NULL,
                            # appearance theme and axis
                            use_theme = TRUE,
                            log_y = FALSE,
@@ -204,8 +206,9 @@ plot.ggeffects <- function(x,
   xif <- attr(x, "x.is.factor", exact = TRUE)
   x_is_factor <- !is.null(xif) && xif == "1"
 
-  # set default size for geoms
+  # set default size and shape for geoms
   if (is.null(dot_size)) dot_size <- 2
+  if (is.null(dot_shape)) dot_shape <- 16
   if (is.null(line_size)) line_size <- 0.7
 
   if (!missing(grid)) facets <- grid
@@ -361,6 +364,7 @@ plot.ggeffects <- function(x,
     show_ci = show_ci,
     ci_style = ci_style,
     dot_size = dot_size,
+    dot_shape = dot_shape,
     line_size = line_size,
     connect_lines = connect_lines,
     case = case,
@@ -434,11 +438,12 @@ plot.ggeffects <- function(x,
 
 plot_panel <- function(x, colors, has_groups, facets_grp, facets, facet_polr,
                        is_black_white, x_is_factor, alpha, dot_alpha, dodge,
-                       show_ci, ci_style, dot_size, line_size, connect_lines,
-                       case, jitter, jitter.miss, show_data, label.data,
-                       residuals, residuals.line, show_title, show_x_title,
-                       show_y_title, show_legend, log_y, y.breaks, y.limits,
-                       use_theme, n_rows = NULL, latent_thresholds, verbose = TRUE, ...) {
+                       show_ci, ci_style, dot_size, dot_shape = NULL, line_size,
+                       connect_lines, case, jitter, jitter.miss, show_data,
+                       label.data, residuals, residuals.line, show_title,
+                       show_x_title, show_y_title, show_legend, log_y, y.breaks,
+                       y.limits, use_theme, n_rows = NULL, latent_thresholds,
+                       verbose = TRUE, ...) {
   # for plotting, we need to convert groups/facets into factors
   if (.obj_has_name(x, "group") && is.character(x$group)) {
     x$group <- factor(x$group, levels = unique(x$group))
@@ -522,8 +527,8 @@ plot_panel <- function(x, colors, has_groups, facets_grp, facets, facet_polr,
   rawdat <- attr(x, "rawdata", exact = TRUE)
   if (show_data) {
     p <- .add_raw_data_to_plot(
-      p, x, rawdat, label.data, ci_style, dot_alpha, dot_size, dodge, jitter,
-      jitter.miss, colors, verbose = verbose
+      p, x, rawdat, label.data, ci_style, dot_alpha, dot_size, dot_shape,
+      dodge, jitter, jitter.miss, colors, verbose = verbose
     )
   }
 
@@ -535,7 +540,7 @@ plot_panel <- function(x, colors, has_groups, facets_grp, facets, facet_polr,
   if (isTRUE(residuals)) {
     p <- .add_residuals_to_plot(
       p, x, residual_data, residuals.line, ci_style, line_size, dot_alpha,
-      dot_size, dodge, jitter, colors, x_is_factor, verbose = verbose
+      dot_size, dot_shape, dodge, jitter, colors, x_is_factor, verbose = verbose
     )
   }
 
@@ -546,8 +551,8 @@ plot_panel <- function(x, colors, has_groups, facets_grp, facets, facet_polr,
   random_effects_data <- attr(x, "random_effects_data", exact = TRUE)
   if (!is.null(random_effects_data)) {
     p <- .add_re_data_to_plot(
-      p, x, random_effects_data, dot_alpha, dot_size, dodge, jitter,
-      verbose = verbose
+      p, x, random_effects_data, dot_alpha, dot_size, dot_shape, dodge,
+      jitter, verbose = verbose
     )
   }
 
@@ -1094,8 +1099,8 @@ plot.see_equivalence_test_ggeffects <- function(x,
 
 #' @keywords internal
 .add_raw_data_to_plot <- function(p, x, rawdat, label.data, ci_style, dot_alpha,
-                                  dot_size, dodge, jitter, jitter.miss, colors,
-                                  verbose = TRUE) {
+                                  dot_size, dot_shape, dodge, jitter, jitter.miss,
+                                  colors, verbose = TRUE) {
   insight::check_if_installed("ggplot2", reason = "to produce plots of adjusted predictions")
 
   # we need an own aes for this
@@ -1174,7 +1179,7 @@ plot.see_equivalence_test_ggeffects <- function(x,
       params = list(
         alpha = dot_alpha,
         size = dot_size,
-        shape = 16
+        shape = dot_shape
       )
     )
 
@@ -1246,8 +1251,9 @@ plot.see_equivalence_test_ggeffects <- function(x,
 
 #' @keywords internal
 .add_residuals_to_plot <- function(p, x, residuals, residuals.line, ci_style,
-                                   line_size, dot_alpha, dot_size, dodge,
-                                   jitter, colors, x_is_factor, verbose = TRUE) {
+                                   line_size, dot_alpha, dot_size, dot_shape,
+                                   dodge, jitter, colors, x_is_factor,
+                                   verbose = TRUE) {
   insight::check_if_installed("ggplot2", reason = "to produce plots of adjusted predictions")
 
   if (!is.null(residuals)) {
@@ -1310,7 +1316,7 @@ plot.see_equivalence_test_ggeffects <- function(x,
       data = residuals,
       show.legend = FALSE,
       inherit.aes = FALSE,
-      params = list(size = dot_size, alpha = dot_alpha, shape = 16)
+      params = list(size = dot_size, alpha = dot_alpha, shape = dot_shape)
     )
 
     if (is.null(jitter) && verbose) {
@@ -1349,7 +1355,7 @@ plot.see_equivalence_test_ggeffects <- function(x,
 
 #' @keywords internal
 .add_re_data_to_plot <- function(p, x, random_effects_data, dot_alpha, dot_size,
-                                 dodge, jitter, verbose = TRUE) {
+                                 dot_shape, dodge, jitter, verbose = TRUE) {
   insight::check_if_installed("ggplot2", reason = "to produce plots of adjusted predictions")
 
   # make sure x on x-axis is on same scale
@@ -1379,7 +1385,7 @@ plot.see_equivalence_test_ggeffects <- function(x,
     data = random_effects_data,
     show.legend = FALSE,
     inherit.aes = FALSE,
-    params = list(size = dot_size, alpha = dot_alpha, shape = 16)
+    params = list(size = dot_size, alpha = dot_alpha, shape = dot_shape)
   )
 
   if (is.null(jitter) && verbose) {