+++ NEWS.md
@@ -30,6 +30,9 @@
 * Added a `get_preditions()` method, which can be used to implement own S3-classes
   to add support for new models to **ggeffects**.
 
+* The `plot()` method gets a `dot_shape` argument, to change the shape of data
+  points when `show_data = TRUE`.
+
 * `test_predictions` gains a `test_args` argument, to optionally pass further
   options to `test` for *emmeans* engine/options.